+++ click-service/src/main/java/clickme/clickme/ClickMeApplication.java
@@ -1,0 +1,15 @@
+package clickme.clickme;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@EnableScheduling
+@SpringBootApplication
+public class ClickMeApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ClickMeApplication.class, args);
+	}
+
+}
