+++ click-service/src/main/java/clickme/clickme/config/HeartRepositoryConfig.java
@@ -1,7 +1,7 @@
 package clickme.clickme.config;
 
-import clickme.clickme.repository.HeartMemoryRepository;
-import clickme.clickme.repository.HeartRepository;
+import clickme.clickme.ranking.domain.HeartMemoryRepository;
+import clickme.clickme.ranking.domain.HeartRepository;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
