--- click-service/src/main/java/clickme/clickme/controller/api/response/RankingResponse.java
@@ -1,4 +1,0 @@
-package clickme.clickme.controller.api.response;
-
-public record RankingResponse(Long ranking, String nickname, Long count, String profileImage) {
-}
