--- click-service/src/main/java/clickme/clickme/domain/AbstractRootEntity.java
@@ -1,22 +1,0 @@
-package clickme.clickme.domain;
-
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.ZonedDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(value = AuditingEntityListener.class)
-public class AbstractRootEntity {
-
-    @CreationTimestamp
-    private ZonedDateTime createdAt;
-
-    @UpdateTimestamp
-    private ZonedDateTime updatedAt;
-}
