+++ click-service/src/main/java/clickme/clickme/ranking/application/RankingService.java
@@ -1,0 +1,31 @@
+package clickme.clickme.ranking.application;
+
+import clickme.clickme.ranking.domain.HeartRepository;
+import clickme.clickme.ranking.domain.MemberRepository;
+import clickme.clickme.ranking.ui.response.RankingResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class RankingService {
+
+    private final HeartRepository heartRepository;
+    private final MemberRepository memberRepository;
+
+    public Long findRankByName(final String name) {
+        return heartRepository.findRankByName(name);
+    }
+
+    public List<RankingResponse> findLiveRanking(final int start, final int end) {
+        return heartRepository.findLiveRanking(start, end)
+                .stream()
+                .map(ranking -> {
+                    String profileImage = memberRepository.getProfileImageUrlByName(ranking.nickname());
+                    return new RankingResponse(ranking.ranking(), ranking.nickname(), ranking.count(), profileImage);
+                })
+                .toList();
+    }
+}
