+++ click-service/src/main/java/clickme/clickme/ranking/domain/HeartRepository.java
@@ -1,0 +1,20 @@
+package clickme.clickme.ranking.domain;
+
+import clickme.clickme.ranking.infrastructure.dto.RankingDto;
+
+import java.util.List;
+
+public interface HeartRepository {
+
+    void increaseCount(final String name);
+
+    void add(final String name);
+
+    void saveChanged(final String name);
+
+    Long findByName(final String name);
+
+    Long findRankByName(final String name);
+
+    List<RankingDto> findLiveRanking(final int start, final int end);
+}
