+++ click-service/src/main/java/clickme/clickme/ranking/domain/exception/NotFoundMemberException.java
@@ -1,0 +1,8 @@
+package clickme.clickme.ranking.domain.exception;
+
+public class NotFoundMemberException extends RuntimeException {
+
+    public NotFoundMemberException(final String name) {
+        super("존재하지 않는 사용자입니다. name = %s".formatted(name));
+    }
+}
