+++ click-service/src/main/java/clickme/clickme/ranking/infrastructure/dto/RankingDto.java
@@ -1,0 +1,4 @@
+package clickme.clickme.ranking.infrastructure.dto;
+
+public record RankingDto(Long ranking, String nickname, Long count) {
+}
