+++ click-service/src/main/java/clickme/clickme/ranking/ui/response/RankingResponse.java
@@ -1,0 +1,4 @@
+package clickme.clickme.ranking.ui.response;
+
+public record RankingResponse(Long ranking, String nickname, Long count, String profileImage) {
+}
