--- click-service/src/main/java/clickme/clickme/repository/dto/RankingDto.java
@@ -1,4 +1,0 @@
-package clickme.clickme.repository.dto;
-
-public record RankingDto(Long ranking, String nickname, Long count) {
-}
