--- click-service/src/main/java/clickme/clickme/repository/exception/NotFoundMemberException.java
@@ -1,8 +1,0 @@
-package clickme.clickme.repository.exception;
-
-public class NotFoundMemberException extends RuntimeException {
-
-    public NotFoundMemberException(final String name) {
-        super("존재하지 않는 사용자입니다. name = %s".formatted(name));
-    }
-}
