+++ click-service/src/main/java/clickme/clickme/svg/domain/count/Count.java
@@ -1,0 +1,37 @@
+package clickme.clickme.svg.domain.count;
+
+public class Count {
+
+    private static final String MAX_COUNT = "99999+";
+    private static final long MAX_COUNT_VALUE = 99999L;
+
+    private final Long number;
+
+    public Count(final Long number) {
+        this.number = number;
+    }
+
+    public boolean isEven() {
+        return this.number % 2 == 0;
+    }
+
+    public boolean isZero() {
+        return this.number == 0;
+    }
+
+    public Count increase() {
+        return new Count(this.number + 1);
+    }
+
+    public String getValue() {
+        if (this.number >= MAX_COUNT_VALUE) {
+            return MAX_COUNT;
+        }
+        return String.valueOf(this.number);
+    }
+
+    public int getLength() {
+        return String.valueOf(this.number)
+                .length();
+    }
+}
