+++ click-service/src/main/java/clickme/clickme/svg/util/EmojiRandomIndexGenerator.java
@@ -1,0 +1,16 @@
+package clickme.clickme.svg.util;
+
+import org.springframework.stereotype.Component;
+
+import java.util.Random;
+
+@Component
+public class EmojiRandomIndexGenerator implements RandomNumberGenerator {
+
+    private static final Random random = new Random();
+
+    @Override
+    public int generator(final int size) {
+        return random.nextInt(size);
+    }
+}
