+++ click-service/src/main/java/clickme/clickme/svg/util/RandomNumberGenerator.java
@@ -1,0 +1,6 @@
+package clickme.clickme.svg.util;
+
+public interface RandomNumberGenerator {
+
+    int generator(final int size);
+}
