--- click-service/src/main/java/clickme/clickme/util/ElementManipulator.java
@@ -1,14 +1,0 @@
-package clickme.clickme.util;
-
-import org.w3c.dom.Element;
-
-public abstract class ElementManipulator {
-
-    protected void setAttribute(Element element,
-                                final String attributeName,
-                                final String value) {
-        element.setAttributeNS(null,
-                attributeName,
-                value);
-    }
-}
