--- click-service/src/main/java/clickme/clickme/util/RandomNumberGenerator.java
@@ -1,6 +1,0 @@
-package clickme.clickme.util;
-
-public interface RandomNumberGenerator {
-
-    int generator(final int size);
-}
