+++ click-service/src/test/java/clickme/clickme/svg/util/RandomNumberGeneratorTest.java
@@ -1,0 +1,19 @@
+package clickme.clickme.svg.util;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class RandomNumberGeneratorTest {
+
+    @Test
+    @DisplayName("랜덤으로 번호를 생성한다.")
+    void createRandomNumber() {
+        RandomNumberGenerator randomNumberGenerator = new EmojiRandomIndexGenerator();
+
+        final int num = randomNumberGenerator.generator(3);
+
+        assertThat(num).isLessThan(3);
+    }
+}
